修改接口 & 实现 getCharCount

+++ 041801114/src/info/tozzger/demo/CounterImpl.java
@@ -24,7 +24,7 @@ public interface CounterImpl {
      * 
      * @return 字符数
      */
-    int getCharCount();
+    long getCharCount();
 
     /**
      *
